# src/main/resources/static/api-docs.yaml
openapi: 3.1.0
info:
  title: Cinema API
  description: REST API для управления кинотеатром — фильмы, сеансы, залы, билеты, жанры и посетители.
  version: v1.0.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
paths:
...
  /tickets:
    get:
      tags: [tickets]
      summary: Получить все билеты (с пагинацией)
      description: Возвращает постраничный список всех билетов. По умолчанию размер страницы — 3.
      operationId: getAllTickets
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: Список билетов успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseTicketDTO"
              example:
                content:
                  - id: 1
                    seatNumber: 15
                    visitorEmail: ivan@example.com
                    session:
                      id: 1
                      startTime: "2025-06-15T18:00:00"
                      movieTitle: Интерстеллар
                      hallName: Красный зал
                      price: 350.0
                metadata:
                  page: 0
                  totalPage: 4
                  size: 3
                  totalElements: 12
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /tickets/{id}:
    get:
      tags: [tickets]
      summary: Получить билет по ID (эффективный запрос)
      description: >
        Возвращает билет по ID с использованием JOIN FETCH — загружает все связанные
        сущности одним запросом, избегая проблемы N+1.
      operationId: getTicketById
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Билет найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDTO"
              example:
                id: 1
                seatNumber: 15
                visitorEmail: ivan@example.com
                session:
                  id: 1
                  startTime: "2025-06-15T18:00:00"
                  movieTitle: Интерстеллар
                  hallName: Красный зал
                  price: 350.0
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /tickets/lazy/{id}:
    get:
      tags: [tickets]
      summary: Получить билет по ID (с ленивой загрузкой — демо N+1)
      description: >
        Возвращает билет по ID с ленивой загрузкой связанных сущностей.
        Демонстрирует проблему N+1 запросов — используйте /tickets/{id} в продакшне.
      operationId: getTicketLazyById
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Билет найден (с дополнительными запросами к БД)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDTO"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /tickets/search:
    get:
      tags: [tickets]
      summary: Поиск билетов по имени посетителя (Native SQL)
      description: Возвращает постраничный список билетов указанного посетителя. Использует нативный SQL.
      operationId: findTicketsByVisitorNative
      parameters:
        - name: visitorName
          in: query
          required: true
          description: Имя посетителя (точное совпадение)
          schema:
            type: string
            example: Иван
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: Результаты поиска
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseTicketDTO"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /tickets/jpqlsearch:
    get:
      tags: [tickets]
      summary: Поиск билетов по имени посетителя (JPQL)
      description: Аналог /tickets/search, но использует JPQL-запрос вместо нативного SQL.
      operationId: findTicketsByVisitorJPQL
      parameters:
        - name: visitorName
          in: query
          required: true
          description: Имя посетителя (точное совпадение)
          schema:
            type: string
            example: Иван
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: Результаты поиска
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseTicketDTO"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /tickets/post:
    post:
      tags: [tickets]
      summary: Создать билет
      description: >
        Создаёт билет для посетителя на указанный сеанс.
        Номер места не должен превышать количество мест в зале.
      operationId: createTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketRequestDTO"
            example:
              sessionId: 1
              visitorId: 2
              seatNumber: 15
      responses:
        "200":
          description: Билет успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDTO"
              example:
                id: 7
                seatNumber: 15
                visitorEmail: ivan@example.com
                session:
                  id: 1
                  startTime: "2025-06-15T18:00:00"
                  movieTitle: Интерстеллар
                  hallName: Красный зал
                  price: 350.0
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /tickets/update/{id}:
    put:
      tags: [tickets]
      summary: Обновить билет
      description: Обновляет данные билета по ID — номер места, сеанс и посетителя.
      operationId: updateTicket
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketRequestDTO"
            example:
              sessionId: 2
              visitorId: 2
              seatNumber: 20
      responses:
        "200":
          description: Билет успешно обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDTO"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /tickets/delete/{id}:
    delete:
      tags: [tickets]
      summary: Удалить билет
      description: Удаляет билет по ID и очищает кэш.
      operationId: deleteTicket
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Билет успешно удалён
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ─── VISITORS ──────────────────────────────────────────────────────────────
  /visitors:
    get:
      tags: [visitors]
      summary: Получить всех посетителей
      description: Возвращает список всех посетителей без пагинации.
      operationId: getAllVisitors
      responses:
        "200":
          description: Список посетителей успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VisitorDTO"
              example:
                - id: 1
                  name: Иван
                  email: ivan@example.com
                - id: 2
                  name: Мария
                  email: maria@example.com
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /visitors/{id}:
    get:
      tags: [visitors]
      summary: Получить посетителя по ID
      description: Возвращает посетителя по указанному идентификатору.
      operationId: getVisitorById
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Посетитель найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitorDTO"
              example:
                id: 1
                name: Иван
                email: ivan@example.com
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /visitors/post:
    post:
      tags: [visitors]
      summary: Создать посетителя
      description: >
        Регистрирует нового посетителя. Имя — от 1 до 20 символов.
        Email должен быть валидным и содержать от 11 до 30 символов.
      operationId: createVisitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitorRequestDTO"
            example:
              name: Алексей
              email: alexey@example.com
      responses:
        "200":
          description: Посетитель успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitorDTO"
              example:
                id: 3
                name: Алексей
                email: alexey@example.com
        "400":
          $ref: "#/components/responses/BadRequest"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/DataIntegrityError"
  /visitors/update/{id}:
    put:
      tags: [visitors]
      summary: Обновить посетителя
      description: Обновляет имя и email посетителя по ID.
      operationId: updateVisitor
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitorRequestDTO"
            example:
              name: Алексей Новый
              email: alexey.new@example.com
      responses:
        "200":
          description: Посетитель успешно обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitorDTO"
              example:
                id: 3
                name: Алексей Новый
                email: alexey.new@example.com
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /visitors/delete/{id}:
    delete:
      tags: [visitors]
      summary: Удалить посетителя
      description: Удаляет посетителя по ID.
      operationId: deleteVisitor
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Посетитель успешно удалён
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"

# ─── COMPONENTS ──────────────────────────────────────────────────────────────
components:
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      description: Уникальный идентификатор ресурса
      schema:
        type: integer
        format: int64
        example: 1
    PageParam:
      name: page
      in: query
      required: false
      description: Номер страницы (начиная с 0)
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
        example: 0
    SizeParam:
      name: size
      in: query
      required: false
      description: Количество элементов на странице (по умолчанию 3)
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 3
        example: 3
    SortParam:
      name: sort
      in: query
      required: false
      description: "Сортировка в формате: поле,направление (например: id,asc)"
      schema:
        type: array
        items:
          type: string
        example: ["id,asc"]
  responses:
    BadRequest:
      description: >
        Некорректный запрос (400). Возможные причины:
        ошибка валидации полей (`@Valid`),
        неверный формат JSON,
        неправильный тип параметра,
        нарушение бизнес-логики (например, номер места превышает вместимость зала).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validationError:
              summary: Ошибка валидации входных данных
              value:
                status: 400
                message: "Некорректный запрос: ошибка валидации входных данных"
                time: "2025-06-15T12:00:00"
            typeMismatch:
              summary: Неправильный тип параметра
              value:
                status: 400
                message: "Некорректный запрос: неправильный тип передаваемого параметра"
                time: "2025-06-15T12:00:00"
            invalidJson:
              summary: Ошибка формата JSON
              value:
                status: 400
                message: "Некорректный запрос: ошибка в формате JSON-сообщения"
                time: "2025-06-15T12:00:00"
            businessLogic:
              summary: Нарушение бизнес-логики (IllegalArgumentException)
              value:
                status: 400
                message: "Некорректный запрос: в зале Красный зал недостаточно мест"
                time: "2025-06-15T12:00:00"
    NotFound:
      description: Ресурс не найден (404). Возникает при обращении к несуществующей сущности.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 404
            message: "Фильм с ID 99 не найден"
            time: "2025-06-15T12:00:00"
    MethodNotAllowed:
      description: >
        Неправильный HTTP-метод (405).
        Возникает при обращении к эндпоинту с недопустимым методом
        (например, POST вместо GET).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 405
            message: "Неправильный тип запроса"
            time: "2025-06-15T12:00:00"
    DataIntegrityError:
      description: >
        Нарушение целостности данных (500).
        Возникает при нарушении ограничений БД — уникальности, внешних ключей и т.п.
        (`DataIntegrityViolationException`, `ConstraintViolationException`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 500
            message: "Нарушение целостности входных данных"
            time: "2025-06-15T12:00:00"
    InternalServerError:
      description: >
        Непредвиденная ошибка сервера (500).
        Возникает при любом необработанном исключении.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 500
            message: "Непредвиденная ошибка на сервере"
            time: "2025-06-15T12:00:00"
  schemas:
...
    TicketRequestDTO:
      type: object
      description: Тело запроса для создания или обновления билета
      required: [sessionId, visitorId, seatNumber]
      properties:
        sessionId:
          type: integer
          format: int64
          description: ID сеанса, на который приобретается билет
          example: 1
        visitorId:
          type: integer
          format: int64
          description: ID посетителя, покупающего билет
          example: 2
        seatNumber:
          type: integer
          format: int32
          description: Номер места (не должен превышать вместимость зала)
          exclusiveMinimum: 0
          example: 15
    VisitorRequestDTO:
      type: object
      description: Тело запроса для создания или обновления посетителя
      required: [name, email]
      properties:
        name:
          type: string
          description: Имя посетителя (от 1 до 20 символов)
          minLength: 1
          maxLength: 20
          example: Иван
        email:
          type: string
          format: email
          description: Email посетителя (от 11 до 30 символов, валидный формат)
          minLength: 11
          maxLength: 30
          example: ivan@example.com

    # ── Response DTOs ─────────────────────────────────────────────────────────
...
    SessionDTO:
      type: object
      description: Данные сеанса
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор сеанса
          example: 1
        startTime:
          type: string
          format: date-time
          description: Время начала сеанса
          example: "2025-06-15T18:00:00"
        movieTitle:
          type: string
          description: Название фильма
          example: Интерстеллар
        hallName:
          type: string
          description: Название зала
          example: Красный зал
        price:
          type: number
          format: double
          description: Цена билета (берётся из зала)
          example: 350.0
    TicketDTO:
      type: object
      description: Данные билета
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор билета
          example: 1
        seatNumber:
          type: integer
          format: int32
          description: Номер места
          example: 15
        session:
          $ref: "#/components/schemas/SessionDTO"
        visitorEmail:
          type: string
          description: Email посетителя, которому принадлежит билет
          example: ivan@example.com
    VisitorDTO:
      type: object
      description: Данные посетителя
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор посетителя
          example: 1
        name:
          type: string
          description: Имя посетителя
          example: Иван
        email:
          type: string
          description: Email посетителя
          example: ivan@example.com

    # ── Pagination ────────────────────────────────────────────────────────────
    Metadata:
      type: object
      description: Метаданные пагинации
      properties:
        page:
          type: integer
          format: int32
          description: Текущий номер страницы (начиная с 0)
          example: 0
        totalPage:
          type: integer
          format: int32
          description: Общее количество страниц
          example: 5
        size:
          type: integer
          format: int32
          description: Размер страницы
          example: 3
        totalElements:
          type: integer
          format: int64
          description: Общее количество элементов во всех страницах
          example: 14
...
    PageResponseTicketDTO:
      type: object
      description: Постраничный ответ со списком билетов
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/TicketDTO"
        metadata:
          $ref: "#/components/schemas/Metadata"

    # ── Error ─────────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      description: Стандартный ответ при ошибке
      properties:
        status:
          type: integer
          format: int32
          description: HTTP-код ошибки
          example: 404
        message:
          type: string
          description: Человекочитаемое описание ошибки
          example: "Фильм с ID 5 не найден"
        time:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: "2025-06-15T12:00:00"